Add ViT encoder + re-organize folders
