--- stable_diffusion/config.py
@@ -1,3 +1,0 @@
-'''
-CONFIGURATION SETTINGS FOR STABLE DIFFUSION PROJECT
-'''
