--- stable_diffusion/src/predict.py
@@ -1,3 +1,0 @@
-'''
-PREDICTION SCRIPT
-'''
