--- stable_diffusion/src/train.py
@@ -1,3 +1,0 @@
-'''
-TRAINING SCRIPT
-'''
